pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalog-cli"
version = "1.0.0"
description = "AI-First Service Catalog CLI — deterministic + RAG-powered queries over Backstage catalog data"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
]
test = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]
dev = [
    "catalog-cli[web,test]",
    "ruff>=0.4.0",
]

[project.scripts]
catalog-cli = "catalog_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
